Support MariaDB in report scripts by adding --dbscheme argument
Allow the database scheme (mysql or MariaDB) to be specified on the
command line for opendmarc-import, opendmarc-reports, and
opendmarc-expire.

This also requires not including the "port" option in the DBD
connection string if the port number is the default, since MariaDB
does not allow the port number to specified when the host name is
localhost.

This change has recently become necessary because when DBD::mysql
connects to a recent MariaDB server it generates a warning, "WARNING:
MYSQL_OPT_RECONNECT is deprecated and will be removed in a future
version," whereas using DBD::MariaDB instead makes the warning go
away.

This change is also a good idea because although connecting to MariaDB
with DBD::mysql works, it's preferable to use DBD::MariaDB.